script: Move some utility bindings code to script_bindings.
